简单优化
